move MockRedis
